Separate out ws/languages page from gSSP
